# api-specs/openapi.yaml
openapi: 3.0.3
info:
  title: ESG & Environmental Data Management Platform API
  description: |
    REST API for the ESG Environmental Data Management Platform.
    Manages energy, water, waste, GHG emissions, air emissions, sales, recovery,
    and production data across manufacturing sites.
  version: 2026-02-24
  contact:
    name: ESG Platform Team
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api/2026-02-24
    description: Local development
  - url: http://localhost:3000/api/2026-02-24
    description: Docker (via nginx proxy)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /auth/login
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationError:
      type: object
      properties:
        error:
          type: string
          example: Validation error
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    SiteRef:
      type: object
      properties:
        name:
          type: string
        code:
          type: string

    # ── Data Entry Common ──
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [admin, site_user, viewer]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userSites:
          type: array
          items:
            type: object
            properties:
              siteId:
                type: integer
              site:
                $ref: '#/components/schemas/SiteRef'
    CreateUser:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 8
          maxLength: 128
        fullName:
          type: string
          minLength: 1
          maxLength: 150
        role:
          type: string
          enum: [admin, site_user, viewer]
          default: viewer
        isActive:
          type: boolean
          default: true
        siteIds:
          type: array
          items:
            type: integer
    UpdateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          type: string
          enum: [admin, site_user, viewer]
        isActive:
          type: boolean
        password:
          type: string
          minLength: 8
        siteIds:
          type: array
          items:
            type: integer
...
  responses:
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: Duplicate record
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationFailed:
      description: Request body validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List all users (admin only)
      responses:
        '200':
          description: Array of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: Create user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User created
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '409':
          $ref: '#/components/responses/Conflict'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user details (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: Updated
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Users]
      summary: Deactivate user (admin only, soft-delete)
      description: Cannot deactivate your own account.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: User deactivated
        '400':
          description: Cannot deactivate own account
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Audit ──
...